refactor: logo 디렉토리 이동
